Split RenderNode.swift into one-file-per-type (file-org policy)
Addresses CodeRabbit: move the sibling public types out of RenderNode.swift into
ModifierArg.swift, RenderModifier.swift, ActionCommand.swift, ButtonAction.swift,
ReorderSpec.swift. RenderNode.swift keeps RenderNode + its nested Kind. Pure move,
no behavior change; package builds, 34 tests pass.
