+++ Packages/CmuxSwiftRender/Sources/CmuxSwiftRender/ButtonAction.swift
@@ -1,0 +1,9 @@
+/// The captured behavior of a `Button`, evaluated when the button is tapped
+/// by a host runtime.
+public struct ButtonAction: Sendable, Equatable {
+    public let commands: [ActionCommand]
+
+    public init(commands: [ActionCommand]) {
+        self.commands = commands
+    }
+}
